+++ internal/controller/operator/factory/vmauth/vmusers_config.go
@@ -20,6 +20,7 @@ import (
 	"k8s.io/utils/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	vmv1 "github.com/VictoriaMetrics/operator/api/operator/v1"
 	vmv1beta1 "github.com/VictoriaMetrics/operator/api/operator/v1beta1"
 	"github.com/VictoriaMetrics/operator/internal/controller/operator/factory/build"
 	"github.com/VictoriaMetrics/operator/internal/controller/operator/factory/k8stools"
@@ -323,11 +324,37 @@ var crdNameToObject = map[string]objectWithURL{
 	"VMCluster/vmselect":  newClusterWithURL("vmselect"),
 	"VMCluster/vminsert":  newClusterWithURL("vminsert"),
 	"VMCluster/vmstorage": newClusterWithURL("vmstorage"),
+	"VLSingle":            &vmv1.VLSingle{},
+	"VLCluster/vlselect":  newClusterWithURL("vlselect"),
+	"VLCluster/vlinsert":  newClusterWithURL("vlinsert"),
+	"VLCluster/vlstorage": newClusterWithURL("vlstorage"),
+	"VLAgent":             &vmv1.VLAgent{},
 }
 
 // helper interface to restore VMCluster type
 type unwrapObject interface {
 	origin() client.Object
+}
+
+var clusterComponentToURL = map[string]func(obj client.Object) string{
+	"vminsert": func(obj client.Object) string {
+		return obj.(*vmv1beta1.VMCluster).VMInsertURL()
+	},
+	"vmselect": func(obj client.Object) string {
+		return obj.(*vmv1beta1.VMCluster).VMSelectURL()
+	},
+	"vmstorage": func(obj client.Object) string {
+		return obj.(*vmv1beta1.VMCluster).VMStorageURL()
+	},
+	"vlinsert": func(obj client.Object) string {
+		return obj.(*vmv1.VLCluster).SelectURL()
+	},
+	"vlselect": func(obj client.Object) string {
+		return obj.(*vmv1.VLCluster).InsertURL()
+	},
+	"vlstorage": func(obj client.Object) string {
+		return obj.(*vmv1.VLCluster).StorageURL()
+	},
 }
 
 type clusterWithURL struct {
@@ -345,17 +372,13 @@ func (c *clusterWithURL) origin() client.Object {
 	return c.vmc
 }
 
+// AsURL implements AsURL interface
 func (c *clusterWithURL) AsURL() string {
-	switch c.component {
-	case "vmselect":
-		return c.vmc.VMSelectURL()
-	case "vmstorage":
-		return c.vmc.VMStorageURL()
-	case "vminsert":
-		return c.vmc.VMInsertURL()
-	default:
+	builder, ok := clusterComponentToURL[c.component]
+	if !ok {
 		panic(fmt.Sprintf("BUG: not expected component=%q for clusterWithURL object", c.component))
 	}
+	return builder(c.Object)
 }
 
 // fetchCRDRefURLs performs a fetch for CRD objects for vmauth users and returns an url by crd ref key name